refactor(server): extract debuggee tools into src/server/handlers/debuggee-tools.ts
start_debugging, restart_debugging, attach_to_process, detach_from_process
and redefine_classes move off the CallTool switch as ToolHandler functions;
bodies verbatim with `this.` -> `ctx.` and `result = X; break;` ->
`return X;`. Each of the four try/catch cases assigns result as the last
statement of its try and of its catch's converting branch, with the case's
break directly after the try/catch; those assignments become returns, which
is the same control flow.
