+++ src/server/handlers/debuggee-tools.ts
@@ -1,0 +1,220 @@
+/**
+ * Debuggee lifecycle tools: start_debugging, restart_debugging,
+ * attach_to_process, detach_from_process, redefine_classes.
+ */
+import { ErrorCode as McpErrorCode, McpError } from '@modelcontextprotocol/sdk/types.js';
+import type { ToolHandler } from '../tool-context.js';
+
+export const startDebuggingTool: ToolHandler = async (ctx, args) => {
+  if (!args.sessionId || !args.scriptPath) {
+    throw new McpError(McpErrorCode.InvalidParams, 'Missing required parameters');
+  }
+
+  try {
+    if (args.adapterLaunchConfig !== undefined) {
+      const cfg = args.adapterLaunchConfig;
+      if (cfg === null || typeof cfg !== 'object' || Array.isArray(cfg)) {
+        throw new McpError(McpErrorCode.InvalidParams, 'adapterLaunchConfig must be an object when provided');
+      }
+    }
+
+    const intake = ctx.normalizeStartDebuggingArgs(args.dapLaunchArgs, args.breakOnExceptions);
+    const debugResult = await ctx.startDebugging(
+      args.sessionId,
+      args.scriptPath,
+      args.args,
+      intake.dapLaunchArgs,
+      args.dryRunSpawn,
+      args.adapterLaunchConfig,
+      ctx.validateBreakOnExceptions(intake.breakOnExceptions)
+    );
+    const responsePayload: Record<string, unknown> = {
+      success: debugResult.success,
+      state: debugResult.state,
+      message: debugResult.error ? debugResult.error : (debugResult.data as Record<string, unknown>)?.message || `Operation status for ${args.scriptPath}`,
+    };
+    if (debugResult.data) {
+      responsePayload.data = debugResult.data;
+    }
+    // Top-level warning join (set_breakpoint pattern): intake
+    // normalization notes (issue #305) plus any session-manager
+    // warning (unbound function breakpoints, issue #308).
+    const dataWarning = (debugResult.data as { warning?: string } | undefined)?.warning;
+    const startWarnings = [...intake.warnings, dataWarning].filter(Boolean);
+    if (debugResult.success && startWarnings.length > 0) {
+      responsePayload.warning = startWarnings.join('; ');
+    }
+    return { content: [{ type: 'text', text: JSON.stringify(responsePayload) }] };
+  } catch (error) {
+    // Handle session state errors specifically
+    if (error instanceof McpError && 
+        (error.message.includes('terminated') || 
+         error.message.includes('closed') || 
+         (error.message.includes('not found') && error.message.includes('Session')))) {
+      return { content: [{ type: 'text', text: JSON.stringify({ success: false, error: error.message, state: 'stopped' }) }] };
+    } else {
+      // Re-throw all other errors (including file validation errors)
+      throw error;
+    }
+  }
+};
+
+export const restartDebuggingTool: ToolHandler = async (ctx, args) => {
+  if (!args.sessionId) {
+    throw new McpError(McpErrorCode.InvalidParams, 'Missing required parameter: sessionId');
+  }
+  try {
+    const debugResult = await ctx.restartDebugging(args.sessionId);
+    const responsePayload: Record<string, unknown> = {
+      success: debugResult.success,
+      state: debugResult.state,
+      message: debugResult.error
+        ? debugResult.error
+        : (debugResult.data as Record<string, unknown>)?.message || 'Debugging restarted',
+    };
+    if (debugResult.error) {
+      responsePayload.error = debugResult.error;
+    }
+    if (debugResult.data) {
+      responsePayload.data = debugResult.data;
+      // Surface the merged restart warning (stale anchors and/or
+      // unbound function breakpoints) at the top level too —
+      // same discoverability as set_breakpoint/start_debugging.
+      const restartWarning = (debugResult.data as { warning?: string }).warning;
+      if (debugResult.success && restartWarning) {
+        responsePayload.warning = restartWarning;
+      }
+    }
+    return { content: [{ type: 'text', text: JSON.stringify(responsePayload) }] };
+  } catch (error) {
+    return ctx.handleBreakpointToolError(error);
+  }
+};
+
+export const attachToProcessTool: ToolHandler = async (ctx, args) => {
+  if (!args.sessionId) {
+    throw new McpError(McpErrorCode.InvalidParams, 'Missing required sessionId');
+  }
+
+  try {
+    if (args.adapterConfig !== undefined) {
+      const cfg = args.adapterConfig;
+      if (cfg === null || typeof cfg !== 'object' || Array.isArray(cfg)) {
+        throw new McpError(McpErrorCode.InvalidParams, 'adapterConfig must be an object when provided');
+      }
+    }
+
+    ctx.logger.info('Attach to process requested', {
+      sessionId: args.sessionId,
+      port: args.port,
+      host: args.host,
+      processId: args.processId
+    });
+
+    const attachResult = await ctx.sessionManager.attachToProcess(args.sessionId, {
+      port: args.port,
+      host: args.host,
+      processId: args.processId,
+      timeout: args.timeout,
+      verifyTimeout: args.verifyTimeout,
+      sourcePaths: args.sourcePaths,
+      stopOnEntry: args.stopOnEntry,
+      justMyCode: args.justMyCode,
+      breakOnExceptions: ctx.validateBreakOnExceptions(args.breakOnExceptions),
+      adapterConfig: args.adapterConfig
+    });
+
+    const responsePayload: Record<string, unknown> = {
+      success: attachResult.success,
+      state: attachResult.state,
+      message: attachResult.error ||
+        (attachResult.data as Record<string, unknown>)?.message ||
+        'Attach operation completed'
+    };
+
+    if (attachResult.data) {
+      responsePayload.data = attachResult.data;
+      // Surface the dropped-adapterConfig-keys warning (issue
+      // #450) at the top level too — same discoverability as
+      // set_breakpoint/start_debugging/restart_debugging.
+      const attachWarning = (attachResult.data as { warning?: string }).warning;
+      if (attachResult.success && attachWarning) {
+        responsePayload.warning = attachWarning;
+      }
+    }
+
+    return { content: [{ type: 'text', text: JSON.stringify(responsePayload) }] };
+  } catch (error) {
+    // Handle session state errors specifically
+    if (error instanceof McpError &&
+        (error.message.includes('terminated') ||
+         error.message.includes('closed') ||
+         (error.message.includes('not found') && error.message.includes('Session')))) {
+      return { content: [{ type: 'text', text: JSON.stringify({
+        success: false,
+        error: error.message,
+        state: 'stopped'
+      }) }] };
+    } else {
+      throw error;
+    }
+  }
+};
+
+export const detachFromProcessTool: ToolHandler = async (ctx, args) => {
+  if (!args.sessionId) {
+    throw new McpError(McpErrorCode.InvalidParams, 'Missing required sessionId');
+  }
+
+  try {
+    ctx.logger.info('Detach from process requested', {
+      sessionId: args.sessionId,
+      terminateProcess: args.terminateProcess
+    });
+
+    const detachResult = await ctx.sessionManager.detachFromProcess(
+      args.sessionId,
+      args.terminateProcess ?? false
+    );
+
+    const responsePayload: Record<string, unknown> = {
+      success: detachResult.success,
+      state: detachResult.state,
+      message: detachResult.error ||
+        (detachResult.data as Record<string, unknown>)?.message ||
+        'Detach operation completed'
+    };
+
+    if (detachResult.data) {
+      responsePayload.data = detachResult.data;
+    }
+
+    return { content: [{ type: 'text', text: JSON.stringify(responsePayload) }] };
+  } catch (error) {
+    // Handle session state errors specifically
+    if (error instanceof McpError &&
+        (error.message.includes('terminated') ||
+         error.message.includes('closed') ||
+         (error.message.includes('not found') && error.message.includes('Session')))) {
+      return { content: [{ type: 'text', text: JSON.stringify({
+        success: false,
+        error: error.message,
+        state: 'stopped'
+      }) }] };
+    } else {
+      throw error;
+    }
+  }
+};
+
+export const redefineClassesTool: ToolHandler = async (ctx, args) => {
+  const redefineResult = await ctx.sessionManager.redefineClasses(
+    args.sessionId as string,
+    args.classesDir as string,
+    (args.sinceTimestamp as number) || 0,
+    args.timeout
+  );
+  return {
+    content: [{ type: 'text' as const, text: JSON.stringify(redefineResult, null, 2) }],
+  };
+};
